Add optimization for various python versions
In python 2, we need to use the low level apis and rewrite the
read/write loop.

In python < 3.7, we need to wrap the file with an iterator to control
the block size.

In python 3.7 we can set the connection's blocksize.

For request, we need to add some ugly hacks; on python 2, the iterator
must be readable. In both versions, we must add a "len" attribute to
make requests use identity transfer encoding and send a content length
header.

This patch separates the ugly requests hack to the requests test module,
and move the other utilities to new util module, since they are not part
of the benchmark infrastructure.