Read and write concurrently
Using pool of buffers, the reader takes a buffer from the pool, read
data into it, and put it on a work queue.

The writer takes queued buffers from the work queue, write them to
storage, and put the buffer back in the pool.

This allows the reader and the writer to wait for I/O concurrently when
using direct I/O.