Update the tests to use the new --workers option
Remove the old parallel_upload helper that started multiple processes,
and use the new --workers option to start multiple worker threads in the
uploaders.

The number of workers is controlled now in the parallel tests, instead
of using the number of cores on a host, and an environment variable.

The go parallel test was removed, will add back when we implement
parallel upload in the go client.